создан компонент который содержит UI компонент Preloader
