+++ src/App.js
@@ -19,8 +19,8 @@ import LessonsPage from './pages/LessonsPage/LessonsPage';
 import LessonsByTeacher from './containers/LessonsByTeacher/LessonsByTeacher';
 import Register from './containers/RegisterContainer/RegisterContainer';
 import Login from './containers/LoginContainer/LoginContainer';
-import Preloader from './components/UI/Preloader/Preloader';
 import studentAttendancePage from './pages/StudentAttendancePage/StudentAttendancePage';
+import MainLoader from './components/MainLoader/MainLoader';
 
 const ProtectedRoute = ({ isAllowed, redirectTo, ...props }) => {
   return isAllowed ? <Route {...props} /> : <Redirect to={redirectTo} />;
@@ -28,22 +28,13 @@ const ProtectedRoute = ({ isAllowed, redirectTo, ...props }) => {
 
 function App() {
   const user = useSelector((state) => state.users.user);
-  const groupsLoading = useSelector((state) => state.groups.loading);
-  const lessonsLoading = useSelector((state) => state.lessons.loading);
-  const paymentsLoading = useSelector((state) => state.payments.loading);
-  const studentsLoading = useSelector((state) => state.students.loading);
-  const subjectsLoading = useSelector((state) => state.subjects.loading);
-  const teachersLoading = useSelector((state) => state.teachers.loading);
-
-  const isOpen =
-    groupsLoading || lessonsLoading || paymentsLoading || studentsLoading || subjectsLoading || teachersLoading;
 
   const isAdminRole = user && user.role === 'admin';
   return (
     <div className='App'>
-      <Preloader isOpen={isOpen} />
       <CssBaseline />
       <Layout>
+        <MainLoader />
         <NotificationContainer />
         <Switch>
           <ProtectedRoute
